Increased performance of ReadString methods by avoiding read unnecessary data from memory
Shortens the size of the buffer used to store the data read from memory in order to avoid reading the entire span when unnecessary.

As of now, reading any string with a total byte size lower than 1024 would still result into reading 1024 bytes. This feels unnecessary as it also results into a (small) performance impact.